@@ -7,7 +7,7 @@
     "bootstrap": "4",
     "commonmark": "^0.28.1",
     "disqus-react": "^1.0.5",
-    "firebase": "^5.5.6",
+    "firebase": "^5.7.0",
     "font-awesome": "^4.7.0",
     "highlight.js": "^9.13.1",
     "i18next": "^12.0.0",
@@ -25,23 +25,23 @@
     "react-redux": "^5.1.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts-ts": "^3.1.0",
+    "react-scripts": "^2.1.1",
     "react-spring": "^6.1.6",
     "react-swipeable": "^4.3.0",
     "react-toastify": "^4.4.0",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
     "speakingurl": "^14.0.1",
-    "three": "^0.98.0"
+    "three": "^0.99.0"
   },
   "scripts": {
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
     "watch-css": "npm run build-css && node-sass-chokidar --include-path ./src --include-path ./node_modules src/ -o src/ --watch --recursive",
-    "start-js": "react-scripts-ts start",
+    "start-js": "react-scripts start",
     "start": "npm-run-all -p watch-css start-js",
     "build-css": "node-sass-chokidar --include-path ./src --include-path ./node_modules src/ -o src/",
-    "build-js": "react-scripts-ts build",
+    "build-js": "react-scripts build",
     "build": "npm-run-all build-css build-js",
     "deploy": "npm run build && firebase deploy",
     "deploy:ci": "npm run build && firebase deploy token $FIREBASE_TOKEN"
@@ -49,20 +49,29 @@
   "devDependencies": {
     "@firebase/app-types": "^0.3.2",
     "@types/i18next": "^11.9.3",
-    "@types/node": "^10.12.1",
-    "@types/react": "^16.4.18",
-    "@types/react-dom": "^16.0.9",
-    "@types/react-redux": "^6.0.9",
+    "@types/jquery": "^3.3.22",
+    "@types/node": "^10.12.12",
+    "@types/react": "^16.7.13",
+    "@types/react-dom": "^16.0.11",
+    "@types/react-redux": "^6.0.11",
     "@types/react-router-dom": "^4.3.1",
     "@types/react-swipeable": "^4.2.2",
     "@types/react-transition-group": "^2.0.14",
-    "@types/three": "^0.93.5",
+    "@types/three": "^0.93.10",
     "babel-core": "^6.26.3",
     "babel-runtime": "^6.26.0",
-    "firebase-tools": "^6.0.0",
+    "firebase-tools": "^6.1.2",
     "node-sass-chokidar": "^1.3.4",
-    "npm-run-all": "^4.1.3",
+    "npm-run-all": "^4.1.5",
     "prop-types": "^15.6.2",
-    "typescript": "^3.1.5"
-  }
+    "tslint": "^5.11.0",
+    "tslint-react": "^3.6.0",
+    "typescript": "^3.2.2"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
