@@ -1,6 +1,6 @@
 import React from 'react';
 import { Redirect } from 'react-router';
-import { LogoutDispatchToProps, LogoutStateToProps } from 'src/containers/admin/logout';
+import { LogoutDispatchToProps, LogoutStateToProps } from '../../containers/admin/logout';
 
 export default class LogoutPage extends React.Component<LogoutStateToProps & LogoutDispatchToProps> {
 
