@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import { ProjectsDispatchToProps, ProjectsStateToProps } from 'src/containers/admin/projects';
-import { ProjectsHome } from 'src/containers/admin/projects/home';
-import { ProjectEditor } from 'src/containers/admin/projects/editor';
+import { ProjectsDispatchToProps, ProjectsStateToProps } from '../../containers/admin/projects';
+import { ProjectsHome } from '../../containers/admin/projects/home';
+import { ProjectEditor } from '../../containers/admin/projects/editor';
 
 type ProjectsPageProps = ProjectsStateToProps & ProjectsDispatchToProps;
 
