@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { Route } from 'react-router';
 import { RouteComponentProps } from 'react-router';
-import { PostsDispatchToProps, PostsStateToProps } from 'src/containers/posts/posts.interfaces';
-import { PostList } from 'src/containers/posts/post-list';
-import { PostPage } from 'src/containers/posts/post-page';
+import { PostsDispatchToProps, PostsStateToProps } from '../../containers/posts/posts.interfaces';
+import { PostList } from '../../containers/posts/post-list';
+import { PostPage } from '../../containers/posts/post-page';
 import './posts.css';
 
 type PostsProps = PostsStateToProps & PostsDispatchToProps & RouteComponentProps<{}>;
