@@ -1,9 +1,9 @@
 import { User } from '@firebase/auth-types';
 import { connect } from 'react-redux';
-import { State } from 'src/redux/reducers';
-import { changeTitle } from 'src/redux/title/actions';
-import { logOut } from 'src/redux/auth/actions';
-import AdminComponent from 'src/components/admin/admin';
+import { State } from '../../redux/reducers';
+import { changeTitle } from '../../redux/title/actions';
+import { logOut } from '../../redux/auth/actions';
+import AdminComponent from '../../components/admin/admin';
 
 export interface AdminStateToProps {
     user: User;
