@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
-import { State } from 'src/redux/reducers';
-import { changeTitle } from 'src/redux/title/actions';
-import { insert, removeError, update } from 'src/redux/database/actions';
-import ProjectEditorComponent from 'src/components/admin/projects/editor';
-import { ProjectInfo } from 'src/containers/admin/projects';
+import { State } from '../../../redux/reducers';
+import { changeTitle } from '../../../redux/title/actions';
+import { insert, removeError, update } from '../../../redux/database/actions';
+import ProjectEditorComponent from '../../../components/admin/projects/editor';
+import { ProjectInfo } from '../projects';
 
 export interface ProjectEditorStateToProps {
     saving: boolean;
@@ -22,9 +22,9 @@ const mapStateToProps = (state: State): ProjectEditorStateToProps => ({
     errorSaving: state.database.errors.projects,
 });
 
-const mapDispatchToProps = (dispatch: any): ProjectEditorDispatchToProps => ({
-    save: (project: ProjectInfo) => dispatch(insert('projects', project)),
-    update: (project: ProjectInfo) => dispatch(update('projects', project, true)),
+const mapDispatchToProps = (dispatch: (...args: any) => void): ProjectEditorDispatchToProps => ({
+    save: (project: Partial<ProjectInfo>) => dispatch(insert('projects', project)),
+    update: (project: Partial<ProjectInfo>) => dispatch(update('projects', project, true)),
     changeTitle: (title: string) => dispatch(changeTitle(title)),
     clearError: () => dispatch(removeError('projects')),
 });
