@@ -1,8 +1,8 @@
 import { connect } from 'react-redux';
-import { translate } from 'react-i18next';
-import { State } from 'src/redux/reducers';
-import { changeTitle } from 'src/redux/title/actions';
-import VizComponent from 'src/components/viz';
+import { withNamespaces } from 'react-i18next';
+import { State } from '../redux/reducers';
+import { changeTitle } from '../redux/title/actions';
+import VizComponent from '../components/viz';
 
 export interface VizStateToProps {
     darkMode: boolean;
@@ -20,5 +20,5 @@ const mapDispatchToProps = (dispatch: any): VizDispatchToProps => ({
     changeTitle: (title: String) => dispatch(changeTitle(title)),
 });
 
-export const Viz = translate()(connect(mapStateToProps, mapDispatchToProps)(VizComponent));
+export const Viz = withNamespaces()(connect(mapStateToProps, mapDispatchToProps)(VizComponent));
 export const Component = Viz;
