@@ -63,7 +63,7 @@ function registerValidSW(swUrl) {
                 // It's the perfect time to display a 'New content is
                 // available; please refresh.' message in your web app.
                 console.log('New content is available; please refresh.');
-                warn('Ey, actualiza la web para obtener los últimos cambios', { autoClose: false });
+                warning('Ey, actualiza la web para obtener los últimos cambios', { autoClose: false });
               } else {
                 // At this point, everything has been precached.
                 // It's the perfect time to display a
